requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
passlib==1.7.4
bcrypt==3.2.0
